backend/requirements.txt
Django==4.2.10
djangorestframework==3.15.1
django-cors-headers==4.3.1
pymongo==4.6.2
dnspython==2.6.1
python-dotenv==1.0.1
Pillow==10.3.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
drf-yasg==1.21.7
channels==4.0.0
